Rockbox as an application: Add an 320x240 SDL application target.
It still works mostly like the simulator. There's also some minor left overs from the sim, but it does not define SIMULATOR.
It installs into the current (build) dir, and you need to run it with '--root .' (because it looks for ./.rockbox and not ./simdisk/rockbox) as options. That's one of the few kludges left that should be resolved soon'ish.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@27310 a1c6a512-1295-4272-9138-f99709370657

+++ tools/buildzip.pl
@@ -474,6 +474,7 @@ $year+=1900;
 # made once for all targets
 sub runone {
     my ($target, $fonts)=@_;
+    my $samedir = abs_path($install) == abs_path($rbdir);
 
     # build a full install .rockbox ($rbdir) directory
     buildzip($target, $fonts);
@@ -498,14 +499,18 @@ sub runone {
         if ($install =~ /\/dev\/null/) {
             die "ERROR: No PREFIX given\n"
         }
-        system("cp -r $rbdir \"$install\" >/dev/null");
+        if (!$samedir) {
+            system("cp -r $rbdir \"$install\" >/dev/null");
+        }
     }
     else {
         system("$ziptool $output $rbdir $target >/dev/null");
     }
 
     # remove the $rbdir afterwards
-    rmtree($rbdir);
+    if (!$samedir) {
+        rmtree($rbdir);
+    }
 };
 
 if(!$exe) {
